refactor: add `which` method to `Worker` trait
Before this change, the `which` crate was used
inside the experiments themselves to lookup the
location of binaries in the $PATH of the target
system.

This worked ok, but left the test suite susceptible
to the $PATH of the machine running the tests.

In this commit, a `which` method is added to the
`Worker` trait. In the real-world `System` impl,
this is a simple call again to the `which` crate.

In the `MockSystem`, provision has been added to
indicate that certain mocked files should be
returned when the `MockSystem::which` function is
called.