fix: use `dpkg-query` to check for installs
Before this commit, `oxidizr` checked if packages
were installed by checking for the word
`installed` in the output of `apt list <pkg>`,
which works fine all the time `apt` is in English,
but not so well if the system uses a different
locale.

This commit instead relies upon `dpkg-query -s`,
which fails if the package is not installed.